Update unit tests (latest roboptim-shared-tests modifications).
There is still room for improvement. I also reindented some of the
unit tests.
