+++ tests/CMakeLists.txt
@@ -27,12 +27,17 @@ ADD_DEFINITIONS(-DIPOPT_LINEAR_SOLVER="${IPOPT_LINEAR_SOLVER}")
 SET(FULL_TEST "OFF"
   CACHE STRING "Whether to run all the (potentially slow) tests or not")
 
-# Generate local-libdir.hh.
-CONFIG_FILES(shared-tests/local-libdir.hh)
+# Include shared test suite.
+SET(PLUGIN_PATH "${CMAKE_BINARY_DIR}/src")
+SET(SOLVER_NAME ${TESTSUITE_SOLVER})
+SET(FUNCTION_TYPE ::roboptim::EigenMatrixDense)
+SET(PROGRAM_SUFFIX "")
+SET(COST_FUNCTION_TYPE ::roboptim::GenericDifferentiableFunction)
+SET(CONSTRAINT_TYPE_1 ::roboptim::GenericLinearFunction)
+SET(CONSTRAINT_TYPE_2 ::roboptim::GenericDifferentiableFunction)
 
 # Make sure local headers are found
 INCLUDE_DIRECTORIES(${CMAKE_BINARY_DIR}/tests)
-
 
 # Add Boost path to include directories.
 INCLUDE_DIRECTORIES(${Boost_INCLUDE_DIRS})
@@ -43,7 +48,7 @@ ADD_DEFINITIONS(-DBOOST_TEST_DYN_LINK -DBOOST_TEST_MAIN)
 SET(COMMON_HEADERS
   ${HEADERS}
   shared-tests/common.hh
-  shared-tests/hs071.hh
+  shared-tests/fixture.hh
   )
 
 # ROBOPTIM_TRAJECTORY_TEST(NAME)
@@ -56,6 +61,10 @@ SET(COMMON_HEADERS
 #
 MACRO(ROBOPTIM_TRAJECTORY_TEST NAME)
   ADD_EXECUTABLE(${NAME} ${NAME}.cc ${COMMON_HEADERS})
+  SET_TARGET_PROPERTIES(${NAME}
+    PROPERTIES COMPILE_FLAGS
+    "-DSOLVER_NAME='\"${SOLVER_NAME}\"' -DPLUGIN_PATH='\"${PLUGIN_PATH}\"' -DFUNCTION_TYPE=${FUNCTION_TYPE} -DCOST_FUNCTION_TYPE=${COST_FUNCTION_TYPE} ${CONSTRAINT_TYPE_1_STR} ${CONSTRAINT_TYPE_2_STR} -DTESTS_DATA_DIR='\"${CMAKE_CURRENT_SOURCE_DIR}\"' -DLOG_FILENAME='\"${NAME}.log\"' ")
+
   ADD_TEST(${NAME} ${RUNTIME_OUTPUT_DIRECTORY}/${NAME})
 
   TARGET_LINK_LIBRARIES(${NAME} roboptim-trajectory)
@@ -76,8 +85,11 @@ ENDMACRO(ROBOPTIM_TRAJECTORY_TEST)
 # Boost and add it to the test suite.
 #
 MACRO(ROBOPTIM_TRAJECTORY_TEST NAME)
-  ADD_EXECUTABLE(${NAME} ${NAME}.cc
-    ${COMMON_HEADERS} anthropomorphic-cost-function.cc)
+  ADD_EXECUTABLE(${NAME} ${NAME}.cc ${COMMON_HEADERS})
+  SET_TARGET_PROPERTIES(${NAME}
+    PROPERTIES COMPILE_FLAGS
+    "-DSOLVER_NAME='\"${SOLVER_NAME}\"' -DPLUGIN_PATH='\"${PLUGIN_PATH}\"' -DFUNCTION_TYPE=${FUNCTION_TYPE} -DCOST_FUNCTION_TYPE=${COST_FUNCTION_TYPE} ${CONSTRAINT_TYPE_1_STR} ${CONSTRAINT_TYPE_2_STR} -DTESTS_DATA_DIR='\"${CMAKE_CURRENT_SOURCE_DIR}\"' -DLOG_FILENAME='\"${NAME}.log\"' ")
+
   ADD_TEST(${NAME} ${RUNTIME_OUTPUT_DIRECTORY}/${NAME})
 
   TARGET_LINK_LIBRARIES(${NAME} roboptim-trajectory)
