+++ tests/anthropomorphic-cost-function-case-6.cc
@@ -20,7 +20,9 @@
 #include "shared-tests/common.hh"
 #include "anthropomorphic-cost-function.hh"
 
-int run_test ()
+BOOST_FIXTURE_TEST_SUITE (trajectory, TestSuiteConfiguration)
+
+BOOST_AUTO_TEST_CASE (trajectory_anthropomorphic_cost_function_case6)
 {
   const double initialX = 0.;
   const double initialY = 0.;
@@ -30,14 +32,14 @@ int run_test ()
   const double finalY = 3.;
   const double finalTheta = M_PI/2.;
 
-  return optimize (initialX,
-		   initialY,
-		   initialTheta,
-		   finalX,
-		   finalY,
-		   finalTheta);
+  int res = optimize (initialX,
+		      initialY,
+		      initialTheta,
+		      finalX,
+		      finalY,
+		      finalTheta);
+
+  BOOST_CHECK (res == 0);
 }
 
-GENERATE_TEST ()
-
-
+BOOST_AUTO_TEST_SUITE_END ()
