+++ tests/anthropomorphic-cost-function.hh
@@ -17,14 +17,193 @@
 
 #ifndef ANTHROPOMORPHIC_COST_FUNCTION_HH
 # define ANTHROPOMORPHIC_COST_FUNCTION_HH
+
+#undef NDEBUG
+
 # include <cmath>
 
-int optimize (double initialX,
+#include <roboptim/trajectory/sys.hh>
+
+#include <fstream>
+
+#include <boost/assign/list_of.hpp>
+#include <boost/mpl/vector.hpp>
+
+#include <roboptim/core/io.hh>
+#include <roboptim/core/finite-difference-gradient.hh>
+#include <roboptim/core/solver-factory.hh>
+
+#include <roboptim/core/visualization/gnuplot.hh>
+#include <roboptim/core/visualization/gnuplot-commands.hh>
+
+#include <roboptim/trajectory/anthropomorphic-cost-function.hh>
+#include <roboptim/trajectory/free-time-trajectory.hh>
+#include <roboptim/trajectory/freeze.hh>
+#include <roboptim/trajectory/frontal-speed.hh>
+#include <roboptim/trajectory/fwd.hh>
+#include <roboptim/trajectory/limit-omega.hh>
+#include <roboptim/trajectory/orthogonal-speed.hh>
+#include <roboptim/trajectory/cubic-b-spline.hh>
+#include <roboptim/trajectory/stable-point-state-function.hh>
+#include <roboptim/trajectory/trajectory-cost.hh>
+
+#include <roboptim/trajectory/visualization/trajectory.hh>
+#include <roboptim/trajectory/visualization/speed.hh>
+
+#include "shared-tests/common.hh"
+
+using namespace roboptim;
+using namespace roboptim::visualization;
+using namespace roboptim::visualization::gnuplot;
+
+typedef Solver<DifferentiableFunction,
+	       boost::mpl::vector<LinearFunction, DifferentiableFunction> >
+test_solver_t;
+
+typedef test_solver_t::problem_t::constraints_t constraint_t;
+typedef FreeTimeTrajectory<CubicBSpline> freeTime_t;
+
+// Problem parameters.
+const unsigned configurationSpaceSize = 3;
+const unsigned nControlPoints = 11;
+
+const unsigned nConstraintsPerCtrlPts = 10;
+const double vMax = 0.4;
+
+
+inline int optimize (double initialX,
 	      double initialY,
 	      double initialTheta,
 	      double finalX,
 	      double finalY,
 	      double finalTheta,
-	      bool setStartingPoint = true);
+	      bool = true)
+{
+  using namespace boost;
+  using namespace boost::assign;
+
+  CubicBSpline::vector_t params (nControlPoints * configurationSpaceSize);
+
+  const double dx = finalX - initialX;
+  const double dy = finalY - initialY;
+  const double dtheta = finalTheta - initialTheta;
+
+  for (unsigned i = 0; i < nControlPoints; ++i)
+    {
+      params[i * configurationSpaceSize] = initialX;
+      params[i * configurationSpaceSize + 1] = initialY;
+      params[i * configurationSpaceSize + 2] = initialTheta;
+
+      params[i * configurationSpaceSize] +=
+	dx / (nControlPoints - 1) * i;
+      params[i * configurationSpaceSize + 1] +=
+	dy / (nControlPoints - 1) * i;
+      params[i * configurationSpaceSize + 2] +=
+	dtheta / (nControlPoints - 1) * i;
+    }
+
+  // Make trajectories.
+  CubicBSpline::interval_t timeRange = CubicBSpline::makeInterval (0., 16.);
+  CubicBSpline spline (timeRange, configurationSpaceSize, params, "before");
+  freeTime_t freeTimeTraj (spline, 1.);
+
+  // Define cost.
+  AnthropomorphicCostFunction<freeTime_t> cost (freeTimeTraj);
+
+  // Create problem.
+  test_solver_t::problem_t problem (cost);
+  problem.startingPoint () = freeTimeTraj.parameters ();
+
+  // Scale has to remain positive.
+  problem.argumentBounds ()[0] = Function::makeLowerInterval (.5);
+
+  const freeTime_t::vector_t freeTimeParams = freeTimeTraj.parameters ();
+
+  std::vector<Function::size_type> indices;
+  indices.push_back (1);
+  indices.push_back (2);
+  indices.push_back (3);
+  indices.push_back (freeTimeParams.size () - 3);
+  indices.push_back (freeTimeParams.size () - 2);
+  indices.push_back (freeTimeParams.size () - 1);
+  makeFreeze (problem) (indices, freeTimeParams);
+
+  // Add constraints on speeds.
+  // Frontal
+  boost::shared_ptr<DerivableFunction> frontalSpeed (new FrontalSpeed ());
+  Function::interval_t vRangeFrontal = Function::makeInterval (0., vMax);
+  StablePointStateFunction<freeTime_t>::addToProblem
+    (freeTimeTraj, frontalSpeed, 1, problem, vRangeFrontal,
+     nControlPoints * nConstraintsPerCtrlPts);
+
+  // Omega (theta dot)
+  Function::interval_t vRangeOmega = Function::makeInterval (-.5, .5);
+  LimitOmega<freeTime_t>::addToProblem
+    (freeTimeTraj, problem, vRangeOmega,
+     nControlPoints * nConstraintsPerCtrlPts);
+
+  // Create the solver plug-in.
+  SolverFactory<test_solver_t> factory (TESTSUITE_SOLVER, problem);
+  test_solver_t& solver = factory ();
+
+  // Set optional log file for debugging
+  SET_LOG_FILE(solver);
+
+  std::cerr << solver << std::endl;
+
+  std::ofstream trajectoryStream ("trajectory.gp");
+  Gnuplot gnuplotTraj = Gnuplot::make_interactive_gnuplot ();
+  gnuplotTraj
+    << set ("multiplot layout 2, 3 title 'trajectory'")
+    << set ("grid")
+    << plot_xy (spline)
+    << plot_xytheta (spline)
+    << plot_speeds (spline);
+
+  test_solver_t::result_t res = solver.minimum ();
+  std::cerr << res << std::endl;
+
+  FreeTimeTrajectory<CubicBSpline> optimizedTrajectory = freeTimeTraj;
+
+  switch (solver.minimumType ())
+    {
+    case GenericSolver::SOLVER_VALUE:
+      {
+	const Result& result = solver.getMinimum<Result> ();
+	optimizedTrajectory.setParameters (result.x);
+	break;
+      }
+
+    case GenericSolver::SOLVER_VALUE_WARNINGS:
+      {
+	const ResultWithWarnings& result =
+	  solver.getMinimum<ResultWithWarnings> ();
+	optimizedTrajectory.setParameters (result.x);
+	break;
+      }
+
+    case GenericSolver::SOLVER_NO_SOLUTION:
+    case GenericSolver::SOLVER_ERROR:
+      return 1;
+    }
+
+  CubicBSpline updatedSpline
+    (optimizedTrajectory.timeRange (),
+     optimizedTrajectory.outputSize (),
+     removeScaleFromParameters (optimizedTrajectory.parameters ()),
+     "after");
+
+  std::cerr
+    << "Final time range: " << updatedSpline.timeRange () << std::endl
+    << "Final length: " << updatedSpline.length () << std::endl;
+
+  gnuplotTraj << plot_xy (updatedSpline)
+	      << plot_xytheta (updatedSpline)
+	      << plot_speeds (updatedSpline)
+	      << unset ("multiplot");
+  trajectoryStream << gnuplotTraj;
+
+  return 0;
+}
 
 #endif //! ANTHROPOMORPHIC_COST_FUNCTION_HH
