+++ tests/state-function.cc
@@ -31,7 +31,9 @@
 using namespace roboptim;
 using namespace roboptim::visualization;
 
-int run_test ()
+BOOST_FIXTURE_TEST_SUITE (trajectory, TestSuiteConfiguration)
+
+BOOST_AUTO_TEST_CASE (trajectory_state_function)
 {
   const unsigned orderMax = 1;
   CubicBSpline::vector_t params (24);
@@ -98,11 +100,10 @@ int run_test ()
       catch (BadGradient<EigenMatrixDense>& bg)
 	{
 	  std::cout << bg << std::endl;
+	  BOOST_CHECK(false);
 	}
       std::cout << "\n\n" << std::endl;
     }
-
-  return 0;
 }
 
-GENERATE_TEST ()
+BOOST_AUTO_TEST_SUITE_END ()
